+++ Changelog.md
@@ -13,6 +13,7 @@ OpenCore Changelog
 - Fixed macserial crashes when processing invalid serials
 - Fixed macserial issues when processing 2021 year serials
 - Added advanced error checking in ocvalidate utility
+- Added `SetupDelay` to configure audio setup delay
 
 #### v0.6.4
 - Added `BlacklistAppleUpdate` to fix macOS 11 broken update optout